Stage one of "empty field at end of line" fix, reworking create_pext_ms and adding unit test to prevent regression once fix is in place.
